% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_thresholds.R
\name{calc_thresholds}
\alias{calc_thresholds}
\title{Calculate threshold values for climate futures}
\usage{
calc_thresholds(SiteID, data = NULL, units = "imperial", directory = tempdir())
}
\arguments{
\item{SiteID}{chosen name to use in file names, attributes, and
directories. (character)}

\item{data}{Default data set to use for .csv creation. Must be created
prior to running function. Follow vignette for example data set creation (data frame)}

\item{units}{the unit type that will be used, defaults to "imperial"
("imperial" or "metric")}

\item{directory}{where to save files to. Per CRAN guidelines, this
defaults to a temporary directory and files created will be lost after
R session ends. Specify a path to retain files.}
}
\value{
one csv file:
SiteID_thresholds.csv
}
\description{
Calculate threshold values for climate futures
}
\examples{
\dontrun{
# Generate sample data

data <- data.frame(
date = sample(seq(as.Date('1950/01/01'), as.Date('2099/12/31'), by="day"), 100),
yr = sample(seq(as.Date('1950/01/01'), as.Date('2099/12/31'), by="year"), 100),
gcm = paste0(rep(letters[1:5], each = 20),
rep(letters[1:20], each = 5),
rep(letters[20:26], each = 1)),
precip = rnorm(100),
tmin = rnorm(100),
tmax = rnorm(100),
rhmax = rnorm(100),
rhmin = rnorm(100),
tavg = rnorm(100)
)

calc_thresholds("SCBL", data = df)
}

}
