% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_tree.R
\name{causal_tree}
\alias{causal_tree}
\title{Trains a single causal tree}
\usage{
causal_tree(
  data,
  target,
  covariates,
  treatment,
  minsize,
  alpha,
  honesty.fraction
)
}
\arguments{
\item{data}{a data frame with predictor, treatment assignment and outcome variables}

\item{target}{name of the outcome variable as character vector}

\item{covariates}{names of the predictors as character vector}

\item{treatment}{name of the treatment indicator variable as character vector (treatment indicator must be binary numeric 1 == treatment, 0 == control with no NA)}

\item{minsize}{minimum number of treatment and control observations that must be present in each split as numeric vector}

\item{alpha}{weighting of cost function numeric vector between 0-1. weights closer to 1 put more emphasis on maximizing heterogeneity. weights closer to 0 put more weight on precisely estimating treatment effects}

\item{honesty.fraction}{fraction of data to be used for honest estimation as numeric vector between 0-1}
}
\value{
trained causal tree (list of three data frames with splitting information)
}
\description{
Trains a single causal tree
}
