% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_pca.R
\name{cf_pca}
\alias{cf_pca}
\title{Calculate climate futures using PCA}
\arguments{
\item{SiteID}{chosen name to use in file names, attributes, and directories. Default
name is "unnamed_site" (character)}

\item{data}{Default dataset to use for the .csv files this function will create.
Follow vignette for example dataset creation. (data frame)}

\item{variables}{Variables you want the PCA to be based off of. Must match column names
from dataframe in \code{data} parameter exactly. If running directly from
\code{summarize_for_pca} and would like to use all threshold values to select models,
write "all_threshold" (character)}

\item{num_cf}{Number of climate futures to select. Option of 2 or 4. Two models will be the
maximum and minimum values of principal component 1 (PC1), and 4 will be the maximum
and minimun values of principal component 2 (PC2)}

\item{directory}{where to save files to. Per CRAN guidelines, this
defaults to a temporary directory and files created will be lost after
R session ends. Specify a path to retain files.}
}
\value{
One (1) csv file that has selected either 2 or 4 models, depending on \code{num_cf} chosen
based upon PCA
One (1) png file of the PCA scatterplot, with models labeled
}
\description{
Designates climate futures of "Warm Wet", "Warm Dry", "Hot Wet", "Hot Dry" and "Central"
and calculates mean values for each GCM. Will additionally calculate which models
represent the largest spread in data of variables entered using PCA. Calculates
summary of threshold values based off of selected summarization parameter and PCA.
}
\details{
*For advanced users only.
}
\examples{

\dontrun{
# Generate sample data

data <- data.frame(
 gcm = c("bcc-csm1-1.rcp45", "BNU-ESM.rcp45", "CanESM2.rcp85", "CCSM4.rcp45",
             "CSIRO-Mk3-6-0.rcp45"),
 precip_change = rnorm(5),
 tmin_change = rnorm(5),
 tmax_change = rnorm(5),
 rhmax_change = rnorm(5),
 rhmin_change = rnorm(5),
 tavg_change = rnorm(5),
 heat_index_change = rnorm(5),
 heat_index_ec_change = rnorm(5),
 heat_index_dan_change = rnorm(5),
 temp_over_95_pctl_change =  rnorm(5),
 temp_over_99_pctl_change =  rnorm(5),
 temp_over_95_pctl_length_change =  rnorm(5),
 temp_under_freeze_change =  rnorm(5),
 temp_under_freeze_length_change =  rnorm(5),
 temp_under_5_pctl_change =  rnorm(5),
 no_precip_change  =  rnorm(5),
 no_precip_length_change =  rnorm(5),
 precip_95_pctl_change =  rnorm(5),
 precip_99_pctl_change =  rnorm(5),
 precip_moderate_change =  rnorm(5),
 precip_heavy_change =  rnorm(5),
 freeze_thaw_change =  rnorm(5),
 gdd_change =  rnorm(5),
 gdd_count_change = rnorm(5),
 not_gdd_count_change = rnorm(5),
 frost_change = rnorm(5),
 grow_length_change = rnorm(5),
 units = rep("imperial", each = 5)
)

cf_pca("SCBL", data = data, variables = c("tmin", "tmax", "rhmin"),
num_cf = 2)
}

}
