% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_function.R
\name{cost_function}
\alias{cost_function}
\title{Causal Tree cost function}
\usage{
cost_function(splits, outcome, treat, minsize, alpha)
}
\arguments{
\item{splits}{logical subsetting vector for observations to use.}

\item{outcome}{name of the outcome variable as character vector.}

\item{treat}{name of the treatment indicator variable as character vector (treatment indicator must be binary numeric 1 == treatment, 0 == control with no NA).}

\item{minsize}{minimum number of treatment and control observations that must be present in each split as numeric vector.}

\item{alpha}{weighting of cost function numeric vector between 0-1. weights closer to 1 put more emphasis on maximizing heterogeneity. weights closer to 0 put more weight on precisely estimating treatment effects.}
}
\value{
numeric vector with split cost
}
\description{
Causal Tree cost function
}
