% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cate.R
\name{get_cate}
\alias{get_cate}
\title{estimate cate in terminal leafs}
\usage{
get_cate(tree, data, treat, outcome)
}
\arguments{
\item{tree}{data.frame with tree splitting structure.}

\item{data}{data.frame of data to estimate cate with.}

\item{treat}{name of the treatment indicator variable as character vector (treatment indicator must be binary numeric 1 == treatment, 0 == control with no NA).}

\item{outcome}{name of the outcome variable as character vector.}
}
\value{
data.frame of terminal leafs with cate estimates.
}
\description{
estimate cate in terminal leafs
}
