% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_causal_forest.R
\name{predict_causal_forest}
\alias{predict_causal_forest}
\title{Estimates conditional average treatment effects based on a trained causal forest and test data}
\usage{
predict_causal_forest(data, cf, predict_oob = TRUE)
}
\arguments{
\item{data}{a data frame with observations to predict and no NA}

\item{cf}{output of the rcf causal_forest function}

\item{predict_oob}{set to TRUE if you want to predict cates on the observations you used to fit the causal forest (uses out of bag observations); set to FALSE if you wish to use a test data set. Default TRUE.}
}
\value{
a data.frame with a cate estimate for each observation.
}
\description{
Estimates conditional average treatment effects based on a trained causal forest and test data
}
