% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcf-package.R
\docType{package}
\name{rcf-package}
\alias{rcf}
\alias{rcf-package}
\title{rcf: Heterogenous Treatment Effect Estimation Using Random Forests}
\description{
rcf is a stripped down experimental rebuild of the grf-labs causal forest algorithm in R and C++. It supports estimation of heterogenous treatment effects and variable importance.
}
\keyword{internal}
