% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcf_data.R
\name{rcf_data}
\alias{rcf_data}
\title{Download data}
\usage{
rcf_data(SiteID, latitude, longitude, units = "imperial")
}
\arguments{
\item{SiteID}{chosen name to use in file names, attributes, and
directories. (character)}

\item{latitude}{latitude of point of interest (spatial)}

\item{longitude}{longitude of point of interest (spatial)}

\item{units}{the unit type that will be used ("imperial" or "metric")}
}
\value{
two dataframes - baseline_all and future_all to be used
with other functions in this package(tibble)
}
\description{
Downloads data from the cft package (https://www.earthdatascience.org/cft/)
to your parent directory.
}
\examples{
rcf_data(SiteID = "SCBL", latitude = 41.83476, longitude =  -103.707)


}
