% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split}
\alias{split}
\title{sample splitting function}
\usage{
split(data, vars, outcome, treat, minsize, alpha)
}
\arguments{
\item{data}{list of predictor, treatment assignment and outcome variables}

\item{vars}{names of the predictors as character vector.}

\item{outcome}{name of the outcome variable as character vector.}

\item{treat}{name of the treatment indicator variable as character vector (treatment indicator must be binary numeric 1 == treatment, 0 == control with no NA).}

\item{minsize}{minimum number of treatment and control observations that must be present in each split as numeric vector.}

\item{alpha}{weighting of cost function numeric vector between 0-1. weights closer to 1 put more emphasis on maximizing heterogeneity. weights closer to 0 put more weight on precisely estimating treatment effects.}
}
\value{
list containing (split = splitting information, subset = logical vector indicating left + right observations, data = observations used in split, var = variable that was split)
}
\description{
sample splitting function
}
