% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{variable_importance}
\alias{variable_importance}
\title{Computes variable importance measures for a causal forest}
\usage{
variable_importance(cf, covariates, n, k)
}
\arguments{
\item{cf}{output of the rcf causal_forest function}

\item{covariates}{names of predictors used in training the causal forest as character vector}

\item{n}{maximum tree depth of splits to be considered for variable importance computation}

\item{k}{decay parameter controlling weighting of splits at different depths for variable importance computation}
}
\value{
a data frame of variables and their importance measures
}
\description{
Computes variable importance measures for a causal forest
}
