% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_thresholds.R
\name{calc_thresholds}
\alias{calc_thresholds}
\title{Calculate threshold values for climate futures}
\usage{
calc_thresholds(
  SiteID = "unnamed_site",
  data = NULL,
  units = "imperial",
  past_years = c(1950, 2000),
  directory = tempdir()
)
}
\arguments{
\item{SiteID}{chosen name to use in file names, attributes, and
directories. (character)}

\item{data}{Default data set to use for .csv creation. Must be created
prior to running function. Follow vignette for example data set creation (data frame)}

\item{units}{the unit type that will be used, defaults to "imperial"
("imperial" or "metric")}

\item{past_years}{years to base past data off of. Cannot be any earlier than 1950.
Must be written as c(past_start, past_end). Defaults to 1950:2000 (numeric)}

\item{directory}{where to save files to. Per CRAN guidelines, this
defaults to a temporary directory and files created will be lost after
R session ends. Specify a path to retain files.}
}
\value{
one csv file:
SiteID_thresholds.csv
}
\description{
Calculate values that exceed or go below a certain number as either "TRUE" of "FALSE."
The summary of these threshold values is calculated in either \code{cf_quadrant} or
\code{summarize_for_pca}functions. This is a very large dataframe and SHOULD NOT be edited
outside of R. Editing this dataframe outside of R will result in a loss of data, and
lead to inaccurate results.
}
\examples{
\dontrun{

# Generate sample data

data <- data.frame(
date = sample(seq(as.Date('1950/01/01'), as.Date('2099/12/31'), by="day"), 100),
yr = rep(c(1960, 1970, 1980, 1990, 2000, 2010, 2020, 2030, 2040, 2050), each = 10),
gcm = rep(c("bcc-csm1-1.rcp45", "BNU-ESM.rcp45", "CanESM2.rcp85", "CCSM4.rcp45",
"CSIRO-Mk3-6-0.rcp45"), each = 20),
precip = rnorm(100),
tmin = rnorm(100),
tmax = rnorm(100),
rhmax = rnorm(100),
rhmin = rnorm(100),
tavg = rnorm(100)
)

calc_thresholds("SCBL", data = df, units = "imperial")
}

}
