% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_forest.R
\name{causal_forest}
\alias{causal_forest}
\title{Trains a causal forest that can be used to estimate conditional average treatment effects.}
\usage{
causal_forest(
  n_trees,
  data,
  outcome,
  covariates,
  treat,
  minsize = 5,
  alpha = 0.05,
  feature_fraction = 0.3,
  sample_fraction = 0.5,
  honest_split = TRUE,
  honesty_fraction = 0.5
)
}
\arguments{
\item{n_trees}{number of trees to grow as numeric vector.}

\item{data}{a data frame with predictor, treatment assignment and outcome variables.}

\item{outcome}{name of the outcome variable as character vector.}

\item{covariates}{names of the predictors as character vector.}

\item{treat}{name of the treatment indicator variable as character vector (treatment indicator must be binary numeric 1 = treatment, 0 = control with no NA).}

\item{minsize}{minimum number of treatment and control observations that must be present in each split as numeric vector. Default 5.}

\item{alpha}{weighting of cost function numeric vector between 0-1. weights closer to 1 put more emphasis on maximizing heterogeneity. weights closer to 0 put more weight on precisely estimating treatment effects. Default 0.05.}

\item{feature_fraction}{fraction of total number of predictors to use in fitting each tree as numeric vector between 0-1. Default 0.3.}

\item{sample_fraction}{fraction of observations to fit each tree with. Default 0.5.}

\item{honest_split}{enables honest splitting. Default TRUE.}

\item{honesty_fraction}{fraction of data to be used for honest estimation as numeric vector between 0-1. Default 0.5.}
}
\value{
list of causal trees.
}
\description{
Trains a causal forest that can be used to estimate conditional average treatment effects.
}
