% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_tree.R
\name{causal_tree}
\alias{causal_tree}
\title{Trains a single causal tree}
\usage{
causal_tree(
  data,
  outcome,
  covariates,
  treat,
  minsize,
  alpha,
  feature_fraction,
  sample_fraction,
  honest_split,
  honesty_fraction
)
}
\arguments{
\item{data}{a data frame with predictor, treatment assignment, outcome variables and no NA.}

\item{outcome}{name of the outcome variable as character vector.}

\item{covariates}{names of the predictors as character vector.}

\item{treat}{name of the treatment indicator variable as character vector (treatment indicator must be binary numeric 1 == treatment, 0 == control with no NA).}

\item{minsize}{minimum number of treatment and control observations that must be present in each split as numeric vector.}

\item{alpha}{weighting of cost function numeric vector between 0-1. weights closer to 1 put more emphasis on maximizing heterogeneity. weights closer to 0 put more weight on precisely estimating treatment effects.}

\item{feature_fraction}{fraction of total number of predictors to use in fitting each tree as numeric vector between 0-1.}

\item{sample_fraction}{fraction of observations to fit tree with.}

\item{honest_split}{enables honest splitting. Default TRUE.}

\item{honesty_fraction}{fraction of data to be used for honest estimation as numeric vector between 0-1.}
}
\value{
trained causal tree (splitting structure, cate estimates, oob samples, data for variable importance computation)
}
\description{
Trains a single causal tree
}
