% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_quadrant.R
\name{cf_quadrant}
\alias{cf_quadrant}
\title{Calculate climate futures based on their quadrant}
\usage{
cf_quadrant(SiteID, data = NULL, year, method = NULL)
}
\arguments{
\item{SiteID}{chosen name to use in file names, attributes, and
directories. (character)}

\item{data}{Default data set to use for .csv creation. Must be created
prior to running function. Follow vignette for example data set creation (data frame)}

\item{year}{year to center changes from historical data around (numeric)}

\item{method}{method for calculating resulting .csv. Options "quadrant" and "corners".
"quadrant" will return a data frame in which all models are labeled
in their respective quadrants "hot wet" "hot dry" "warm wet" "warm dry" and "central
"corners" will do the same, but additionally return the model that represents
the most extreme scenario in each quadrant}
}
\value{
three .csv files:
\enumerate{
\item SiteID_daily_past.csv - past data by model and day, GCM is historical
\item SiteID_future_means_x.csv - future means of all models calculated on central year
3' SiteID_daily_future_x.csv - future data by model and day, future quadrants labeled
_x: c = corner, q = quadrant
}
}
\description{
Calculate climate futures based on their quadrant
}
\examples{
# Generate sample data

df <- data.frame(
date = sample(seq(as.Date('1950/01/01'), as.Date('2099/12/31'), by="day"), 100),
yr = sample(seq(as.Date('1950/01/01'), as.Date('2099/12/31'), by="year"), 100),
GCM = paste0(rep(letters[1:5], each = 20), rep(letters[1:20], each = 5), rep(letters[20:26], each = 1)),
precip = rnorm(100),
tmin = rnorm(100),
tmax = rnorm(100),
rhmax = rnorm(100),
rhmin = rnorm(100),
tavg = rnorm(100)
)

cf_quadrant("BAND", data = df, year = 2040, method = "corners")

}
