% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_quadrant.R
\name{cf_quadrant}
\alias{cf_quadrant}
\title{Calculate climate futures based on their quadrant}
\usage{
cf_quadrant(
  SiteID,
  data = NULL,
  year,
  method = "quadrant",
  summarize_by = "year",
  directory = tempdir()
)
}
\arguments{
\item{SiteID}{chosen name to use in file names, attributes, and
directories. (character)}

\item{data}{Default data set to use for .csv creation. Must be created
prior to running function. Follow vignette for example data set creation (data frame)}

\item{year}{year to center changes from historical data around (numeric)}

\item{method}{method for calculating resulting .csv. Options "quadrant" and "corner".
"quadrant" will return a data frame in which all models are labeled
in their respective quadrants "hot wet" "hot dry" "warm wet" "warm dry" and "central
"corner" will do the same, but additionally return the model that represents
the most extreme scenario in each quadrant}

\item{summarize_by}{how to summarize the data, options are "month", "season", "year"}

\item{directory}{where to save files to. Per CRAN guidelines, this
defaults to a temporary directory and files created will be lost after
R session ends. Specify a path to retain files.}
}
\value{
Two (2) .csv files:
1.SiteID_future_means_x.csv - future means of all models calculated on central year
2. SiteID_future_y_x.csv - Summary of threshold values calculated based on method (x) and
summary metric (y). Means are taken based on month/season/year and either all models in
a quadrant for quadrant method, or most extreme model in each quadrant for corner method
x: c = corner, q = quadrant
y: month, season or year
}
\description{
Calculate climate futures based on their quadrant
}
\examples{
\dontrun{
# Generate sample data

df <- data.frame(
date = sample(seq(as.Date('1950/01/01'), as.Date('2099/12/31'), by="day"), 100),
yr = rep(c(1960, 1970, 1980, 1990, 2000, 2010, 2020, 2030, 2040, 2050), each = 10),
gcm = rep(c("bcc-csm1-1.rcp45", "BNU-ESM.rcp45", "CanESM2.rcp85", "CCSM4.rcp45",
"CSIRO-Mk3-6-0.rcp45"), each = 20),
precip = rnorm(100),
tmin = rnorm(100),
tmax = rnorm(100),
rhmax = rnorm(100),
rhmin = rnorm(100),
tavg = rnorm(100)
)

cf_quadrant("SCBL", data = df, year = 2040, method = "corner", summarize_by = "year")
}
}
