% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_thresholds.R
\name{pca_thresholds}
\alias{pca_thresholds}
\title{Summarize threshold values by PCA and month, season, or year}
\usage{
pca_thresholds(
  SiteID,
  pca_data = NULL,
  all_data = NULL,
  year,
  summarize_by = NULL,
  directory = tempdir()
)
}
\arguments{
\item{SiteID}{chosen name to use in file names, attributes, and
directories. (character)}

\item{pca_data}{The data that results from the \code{cf_pca} function (dataframe)}

\item{all_data}{The data that results from the \code{calc_thresholds} function (dataframe)}

\item{year}{year to center changes from historical data around (numeric)}

\item{summarize_by}{how to summarize the data, options are "month", "season", "year"}

\item{directory}{where to save files to. Per CRAN guidelines, this
defaults to a temporary directory and files created will be lost after
R session ends. Specify a path to retain files.}
}
\value{
One (1) csv file that summarizes threshold values based on PCA and month/season/year
}
\description{
Summarize threshold values by PCA and month, season, or year
}
\examples{
#' @examples
\dontrun{
# Generate sample data

data <- data.frame(
date = sample(seq(as.Date('1950/01/01'), as.Date('2099/12/31'), by="day"), 100),
yr = sample(seq(as.Date('1950/01/01'), as.Date('2099/12/31'), by="year"), 100),
gcm = paste0(rep(letters[1:5], each = 20),
rep(letters[1:20], each = 5),
rep(letters[20:26], each = 1)),
precip = rnorm(100),
tmin = rnorm(100),
tmax = rnorm(100),
rhmax = rnorm(100),
rhmin = rnorm(100),
tavg = rnorm(100)
heat_index = rnorm(100)
heat_index_ec = rnorm(100),
heat_index_dan = rnorm(100),
temp_over_95_pctl =  sample(x = c("TRUE","FALSE"), size = 100, replace = TRUE),
temp_over_99_pctl =  sample(x = c("TRUE","FALSE"), size = 100, replace = TRUE),
temp_over_95_pctl_length =  sample(x = c("TRUE","FALSE"), size = 100, replace = TRUE),
temp_under_freeze =  sample(x = c("TRUE","FALSE"), size = 100, replace = TRUE),
temp_under_freeze_length =  sample(x = c("TRUE","FALSE"), size = 100, replace = TRUE),
temp_under_5_pctl =  sample(x = c("TRUE","FALSE"), size = 100, replace = TRUE),
no_precip  =  sample(x = c("TRUE","FALSE"), size = 100, replace = TRUE),
no_precip_length =  sample(x = c("TRUE","FALSE"), size = 100, replace = TRUE),
precip_95_pctl =  sample(x = c("TRUE","FALSE"), size = 100, replace = TRUE),
precip_99_pctl =  sample(x = c("TRUE","FALSE"), size = 100, replace = TRUE),
precip_moderate =  sample(x = c("TRUE","FALSE"), size = 100, replace = TRUE),
precip_heavy =  sample(x = c("TRUE","FALSE"), size = 100, replace = TRUE),
freeze_thaw =  sample(x = c("TRUE","FALSE"), size = 100, replace = TRUE),
gdd =  sample(x = c("TRUE","FALSE"), size = 100, replace = TRUE),
gdd_count = rnorm(100),
not_gdd_count = rnorm(100),
frost = sample(x = c("TRUE","FALSE"), size = 100, replace = TRUE),
grow_len = rnorm(100)
)

pca_thresholds("SCBL", pca_data = pca_df, all_data = data, year = 2040,
summarize_by = "month", directory = my_project_directory)
}

}
