% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_causal_forest.R
\name{predict_causal_forest}
\alias{predict_causal_forest}
\title{Estimates conditional average treatment effects based on a trained causal forest and test data}
\usage{
predict_causal_forest(cf, test)
}
\arguments{
\item{cf}{output of the rcf causal_forest function}

\item{test}{a data frame of test data}
}
\value{
result list consisting of: 1 data frame with predictors, treatment assignment and cate estimates, 2 list of vectors with cate estimates for each observation
}
\description{
Estimates conditional average treatment effects based on a trained causal forest and test data
}
