% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcf_data.R
\name{rcf_data}
\alias{rcf_data}
\title{Download MACA data from the cft package}
\usage{
rcf_data(
  SiteID,
  latitude,
  longitude,
  units = "imperial",
  directory = tempdir()
)
}
\arguments{
\item{SiteID}{chosen name to use in file names, attributes, and
directories. (character)}

\item{latitude}{latitude of point of interest (spatial)}

\item{longitude}{longitude of point of interest (spatial)}

\item{units}{the unit type that will be used,
defaults to "imperial" ("imperial" or "metric")}

\item{directory}{where to save files to. Per CRAN guidelines, this
defaults to a temporary directory and files created will be lost after
R session ends. Specify a path to retain files.}
}
\value{
one data frame - SiteID.csv - to be used
with other functions in this package(tibble)
}
\description{
Downloads data from the cft package (https://www.earthdatascience.org/cft/)
to your parent directory.
}
\examples{
\dontrun{
rcf_data(SiteID = "SCBL", latitude = 41.83476, longitude =  -103.707)
}

}
